-map rewrite, fixes conn rendering for most conns
